+++ src/Compilers/Lua/Portable/Syntax/InternalSyntax/SyntaxFactory.cs
@@ -121,6 +121,9 @@
         internal static SyntaxToken Literal(GreenNode? leading, string text, long value, GreenNode? trailing) =>
             SyntaxToken.WithValue(SyntaxKind.NumericLiteralToken, leading, text, value, trailing);
 
+        internal static SyntaxToken Literal(GreenNode? leading, string text, ulong value, GreenNode? trailing) =>
+            SyntaxToken.WithValue(SyntaxKind.NumericLiteralToken, leading, text, value, trailing);
+
         internal static SyntaxToken Literal(GreenNode? leading, string text, double value, GreenNode? trailing) =>
             SyntaxToken.WithValue(SyntaxKind.NumericLiteralToken, leading, text, value, trailing);